+++ src/components/ui/terminal/screen/screen.cpp
@@ -526,7 +526,7 @@ namespace qor{ namespace components{ namespace tui {
         {
             if (hyperlinks_[i] == link) 
             {
-                return i;
+                return static_cast<std::uint8_t>(i);
             }
         }
         if (hyperlinks_.size() == std::numeric_limits<std::uint8_t>::max()) 
@@ -534,7 +534,7 @@ namespace qor{ namespace components{ namespace tui {
             return 0;
         }
         hyperlinks_.push_back(std::string(link));
-        return hyperlinks_.size() - 1;
+        return static_cast<std::uint8_t>(hyperlinks_.size() - 1);
     }
 
     const std::string& Screen::Hyperlink(std::uint8_t id) const
